Changes to weapon blaster and spread
For blaster I increased the velocity of the projectile by 5 times to it. with this change it seems to increase the fire rate. while for the spread, I added extra projectiles while delaying the speed of the velocity by half. for this it seems to slow down the rate of fire, but I could be wrong.